[BUGFIX] dynmbox re-implementation
Re-implementation of the dynmbox for Windows and Linux using pthread
locking primitives to ensure read/write atomicity, which the
implementation using pipes could not provide.

This fixes a bug when sending messages grater than PIPE_BUF bytes in
a multi-threaded program, which would cause dynmbox_peek() to report
truncated or corrupted messages, and potentially corrupting memory by
writing beyond the destintion buffer.
This also adds a Windows implementation of the dynmbox.

This new implementation uses a circular buffer to implement the mail box
queue, and standard pthread locking promitives to ensure operation
atomicity, regardless of the message length.

+++ include/futils/dynmbox.h
@@ -45,7 +45,7 @@
 #ifndef __DYNMBOX_H__
 #define __DYNMBOX_H__
 
-#include <unistd.h>
+#include <stdint.h>
 
 #ifdef __cplusplus
 extern "C" {
@@ -54,7 +54,7 @@ extern "C" {
 /* Maximum size of a dynmbox message.
  * This was originally based on Linux default pipe capacity.
  */
-#define DYNMBOX_MAX_SIZE (65536 - sizeof(size_t))
+#define DYNMBOX_MAX_SIZE (65536 - sizeof(uint32_t))
 
 struct dynmbox;
